refactor: moved authentication service to application layer
